backend/auth/requirements.txt
fastapi 
uvicorn 
authlib 
python-jose[cryptography]
pydantic-settings